update:  clean code, solving bugs
