+++ src/index.html
@@ -22,14 +22,14 @@
   <main>
     <ul id="items">
       <li>
-        <img class="item-image" src="https://img.redro.pl/fototapety/spaceship-rocket-space-stars-background-space-background-galaxy-vector-illustration-eps10-700-214110858.jpg" alt="spaceship">
+        <img class="item-image" src="#" alt="spaceship">
         <div class="name-like">
           <p class="name">Space 1</p>
           <i class="like-btn far fa-heart"></i>
           <p class="likes-count">4</p>
         </div>
-        <a href="">Comments</a>
-        <a class="item_detail" href="">Details</a>
+        <a class="item_comment" href="">Comments</a>
+        <a class="item_detail" href="">Reserve</a>
       </li>
     </ul>
   </main>
